Update partition_space.md
While porting the KokkosKernels CI to a new machine, I encountered a hard-coded number of partitions in one of our unit tests. The runtime provided by kokkos indicated that this test was splitting the pool size on OpenMP into too many execution space partitions. It was not clear from the wiki docs how to determine the number of available partitions. This commit is intended to make the documentation clearer.