Change `Experimental::for_each` return type to void
To reflect the changes from kokkos/kokkos#6910

We were returning a copy of the function object argument and now return
nothing to align with `std::for_each` overload that takes an execution
policy.  This is an experimental facility and advertising that we return
`void` is compatible with our policies so we don't need to mention the
change in the doc.