Reorganize options and document BYPASS_ATOMICS and DEPRECATION_WARNINGS
Fix #518

Rework categories
* General options
* Debugging
* Backend-specific options
* Development

to help users navigate options.  In particular, add a note to hints that
dev options are not intended for them to use.

Document `Kokkos_ENABLE_DEPRECATION_WARNINGS` that was missing and
`Kokkos_ENABLE_ATOMICS_BYPASS` that was added in 4.3

Indicate that `Kokkos_ENABLE_DEPRECATED_CODE_3` is removed in 4.3