create_mirror and create_mirror_view: clarified difference
Added a paragraph to "Usage", before the code example, and added comments to the code example.

I was thinking about changing the variable names in the second half of the code example from
`host_mirror_space*` to `space_mirror*`, as now they don't *have* to be in host space, if another space is provided. Though I'm not yet familiar enough with Kokkos to know whether `ExecName()` is pseudo code or evaluates to the default execution space or something like that.